### 2025-11-10 e6d8656
shortened paper, improved readme
