+++ paper/paper.md
@@ -12,6 +12,9 @@ authors:
     surname: Aumedes
     orcid: 0009-0006-9011-0853
     affiliation: 1
+  - given-names: Xavier
+    surname: Celades
+    affiliation: 1
   - given-names: Josep
     surname: Pijuan
     affiliation: 1
@@ -31,32 +34,7 @@ This paper describes the concepts and the techniques used in the development of 
 
 # Statement Of Need 
 WoT-Server was created as a centralised data management solution for the ACQUAOUNT Project. One of the objectives of the ACQUAOUNT Project is providing smart irrigation recommendations to farmers using a water balance model. This model takes as input timeseries data of different properties of the specified field, such as temperature, soil moisture and wind speed, and calculates the best date and amount of irrigation. WoT-Server was developed as a standardized API to allow easy access to faming/water management data for further model implementation and to simplify the upload, storage and retrieval procedures of measurements from sensors in the field. WoT-Server is capable of both receiving data via an HTTP endpoint or fetching the data itself from other API services, a feature useful for centralising data from multiple platforms.  
-Since it was needed that users and developers could interact with the platform easily and consistently, WoT-Server follows the Web-Of-Things standard. This standard provides a common framework for describing and accessing connected devices and services through Thing Descriptions, enabling a uniform interface for interaction regardless of device type or communication protocol. However, since WoT-Server also needed a well-defined data model to store and organise its sensor data, the platform uses the OGC SensorThings standard as the database structure, which provides a standardised model for representing IoT entities such as Things, Sensors, Properties, and Datastreams, along with their timeseries Observations. Using SensorThings ensures that all data stored in WoT-Server follows a consistent and extensible structure, making it easier to query, analyse, and share.  
 All AI-based applications require data, and most if not all IoT applications obtain this data from sensors in timeseries format. The aim of WoT-Server is to provide a platform to easily manage this type of data with minimal setup and configuration required, letting users focus on the tasks specific to their use case. With WoT-Server, all developers creating IoT applications can have a platform where they can store their data up and running in minutes, which they can use to create AI models.
-
-# Definition Of Standards Used
-## WoT
-The Web-Of-Things (WoT) standard is a tool that provides a set of standardized technology building blocks used to simplify the IoT environment for the developers navigating it. The first step of abstraction of WoT is dividing all the interactions of a Thing with the physical world into three different categories, called interaction affordances. A Thing can be anything that is an object of the physical world (any physical thing) or the information world (a virtual thing), and an interaction affordance can be a property, an action, or an event. [@w3c2024]  
-
- + A Property affordance is a value available to be read or written from either the device or the physical world, such as the temperature (which is read from the physical world) or configuration parameter of the device (which can be read or written from another device).  
- + An Action affordance allows other devices to invoke an action by the device, which may or may not interact with the physical world, such as brewing a coffee for a smart coffee machine or starting the process of communicating with another device.  
- + Finally, an Event affordance is used to observe events that can occur in the device, such as a door being open or a cooking machine running out of some ingredient.  
-
-All the metadata of a device and all its interaction affordances are documented in a Thing Description (TD), and something described by a TD is considered a Thing. This file is in JSON-LD format and includes all the data required to use the defined Thing except the network protocol, so that the same TD can be reused over different protocols such as HTTP, MQTT, or others. The protocol is provided using the WoT Binding Template. [@kaebisch2024]  
-
-![Diagram representing how WoT simplifies the interaction between apps and devices.\label{fig1}](diagramawot.png){ width=65% }  
-
-Taking a look at \autoref{fig1}, it is seen how providing a uniform interface for objects facilitates the tasks of the IoT applications, simplifying the connections and creating easily maintainable and expandable applications.   
[…]
