+++ README.md
@@ -1,15 +1,46 @@
-# ACQUAOUNT
-Versió de node **v18.12.0**  
-Per instal·lar:
->npm install
+# ACQUAOUNT WoT-Server
+Node.js version required: **v18.12.0**    
+To run server only locally:
+>npm install  
+>node src/main/server.js src/resources/configs/prod.json
+ 
+To deploy:
+>docker compose up --build -d  
 
-#How to deploy: 
+## Statement Of Need
+WoT-Server was created as a centralised data management solution for the ACQUAOUNT Project. One of the objectives of the ACQUAOUNT Project is providing smart irrigation recommendations to farmers using a water balance model. This model takes as input timeseries data of different properties of the specified field, such as temperature, soil moisture and wind speed, and calculates the best date and amount of irrigation. WoT-Server was developed as a standardized API to allow easy access to faming/water management data for further model implementation and to simplify the upload, storage and retrieval procedures of measurements from sensors in the field. WoT-Server is capable of both receiving data via an HTTP endpoint or fetching the data itself from other API services, a feature useful for centralising data from multiple platforms.  
+Since it was needed that users and developers could interact with the platform easily and consistently, WoT-Server follows the Web-Of-Things standard. This standard provides a common framework for describing and accessing connected devices and services through Thing Descriptions, enabling a uniform interface for interaction regardless of device type or communication protocol. However, since WoT-Server also needed a well-defined data model to store and organise its sensor data, the platform uses the OGC SensorThings standard as the database structure, which provides a standardised model for representing IoT entities such as Things, Sensors, Properties, and Datastreams, along with their timeseries Observations. Using SensorThings ensures that all data stored in WoT-Server follows a consistent and extensible structure, making it easier to query, analyse, and share.  
+All AI-based applications require data, and most if not all IoT applications obtain this data from sensors in timeseries format. The aim of WoT-Server is to provide a platform to easily manage this type of data with minimal setup and configuration required, letting users focus on the tasks specific to their use case. With WoT-Server, all developers creating IoT applications can have a platform where they can store their data up and running in minutes, which they can use to create AI models.
 
-git diff --name-only
-#git add --all
+## Thing Interaction Affordances
 
-!!!!!!!copy all Custom fields from the server to the repostory!!!!
-Custom fields are located in /Projects/ACQUAOUNT/acquaount-platform/src/resources/thingDescription/Fields/Custom/ 
-Download them to your PC and push to the repo (trying to push to git from server fails)
+Each physical location can be interacted with by following the WoT protocol, with the following interaction affordances:  
 
-docker compose up -d --build --force-recreate acquaount-platform
+ + (field/station/item)Information: Provides some static information about the field/station/item of this thing.  
+ + propertiesList: Provides a list of all the properties being observed at this field/station/item.  
+ + propertyInformation: Provides detailed information about one of the properties.  
+ + datastreamsList: Provides a list of datastreams in the field/station/item. A datastream is the flow o data defined by the combination of a Thing, Sensor, and Property.  
+ + datastreamInformation: Provides detailed information about one datastream.  
+ + datastreamLastMeasure: Provides the last measure taken in a specified datastream.  
+ + datastreamMeasures: Provides all the measures taken in a specified datastream.  
+ + lastMeasures: Provides the last measures taken in all datastreams.  
+
+Some types of items also have the following properties:  
+
+ + sensorsList: Provides a list of the sensors in this field/item.  
+ + sensorInformation: Provides detailed information about one of the sensors.  
+
+All types of things have available at least this action affordance, used to upload data to the platform.  
+
+ + receiveMeasure: Used as an endpoint for the sensors to upload data to the platform.
+
[…]
