### 2025-11-10 #9350 @Galahad3x
@danielskatz I have updated both the paper and the README, simplifying the paper and expanding the README, let me know if the changes are satisfactory.
In regards to the commit history, it was all commited at once since the bulk of the development was done in a Git repository stored in a private GitLab instance, which cannot be openly published. I could try to export some kind of log or history of commits, would that work?
