### 2025-11-19 #9470 @editorialbot
[REVIEW]: WoT-Server: Combining the Web-Of-Things and SensorThings standards to build an IoT platform
**Submitting author:** @Galahad3x (<a href="http://orcid.org/0009-0006-9011-0853">Joel Aumedes Serrano</a>)
**Repository:** https://github.com/Applied-Artificial-Intelligence-Eurecat/acquaount
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
