### 2025-12-12 #9470 @zeyus
I have experienced similar issues to @Ahanaf-Aziz , I did get the docker version running, but not without modification. I believe there are some hard-coded URLs and IP addresses in there.

I'm also missing any kind of testing - automated, or even manual instructions. I do have some IoT devices (e.g. ESP32) I could set up to provide data on a closed network.
