### 2026-01-16 #9470 @Galahad3x
Hello all,
I've updated the repository and the paper to fix all the mentioned issues. These are the changes that I've applied:

### Documentation
Rewrote the beginning of the README.md file, adding a Getting Started section that better describes the prerequisites and the installation and start up process.

### Functionality
Moved the base URLs from hardcoded in the Thing Descriptions to a variable in the configuration files.
Fixed the Duplicate Thing bug.

### Testing And Verification
Expanded the Postman collection to include descriptions and examples of successful inputs for every request.

### Paper
Rewrote some sections, added a state of the art section, and performed some grammar and clarity modifications.
Also added the missing references and created a copy of the paper.bib file available in both the repository root and the paper folder.

### Community Guidelines
Added a CONTRIBUTING.md file to describe how contributions to the project should be done.
