### 2026-01-17 #9470 @shahchiragh
Hi @Galahad3x!

Thanks for addressing @Ahanaf-Aziz comments. I had similar issues earlier but so far, the updates look good. I ran the server locally both with and without Docker and it started up cleanly in both cases. E.g.

`[
"[http://192.168.1.242:80/main](http://192.168.1.242/main)",
"[http://192.168.1.242:80/acquaounttelalalthahab](http://192.168.1.242/acquaounttelalalthahab)",
"[http://192.168.1.242:80/acquaountnarc](http://192.168.1.242/acquaountnarc)",
...
...
...
]
`

Overall writing quality is good; the main items left are structural polish and tightening up citations.

Minor suggestions:

- I intentionally left the "Quality of Paper" and "References: sections unchecked because they need improvement. Please add explicit citations where comparisons are made and ensure the reference list matches in-text mentions.
- Run a final pass for grammar, passive voice, and consistency in terminology.

Let me know if you have any questions.

@danielskatz these look like minor edits rather than major rewrites, so we should be close to final review once the references are completed and the structure is polished up for the paper.
