### 2026-01-28 #9470 @Galahad3x
> [@Galahad3x](https://github.com/Galahad3x) - Have you been able to address [@shahchiragh](https://github.com/shahchiragh)'s comments?

Yes, I have just commited now a new version of the document and the references.
