### 2026-02-05 #9470 @zeyus
so, some hiccups with the setup still. From a fresh start, I did get the containers started and the endpoints available which is great.

 1. There is no mention that to run the python script you must have (obviously) python (version?) installed, and that you must run `python3 -m pip install -r src/resources/databaseInitialisation/requirements.txt` , I know this is standard, but I assumed there were no requirements and the script failed the first time.
 2. when I `python3 src/resources/databaseInitialisation/init_database.py src/resources/databaseInitialisation/twosensors.csv  http://127.0.0.1:8008/` (the arguments are not documented, I found them in the source), then it connects but then the output is just:
  ```
  src/resources/databaseInitialisation/init_database.py
  Error in request
  Error in request
  Error in request
  Error in request
  Error in request
  ```

Most of the postman examples run but quite a few fail, and without sufficient documentation, it's hard to know what should be changed to make them work.

maybe @Ahanaf-Aziz has some advice on what they encountered during the postman testing?

I'm not sure how to word this correctly, and maybe it's just revealing my lack of domain-specific knowledge here, but it seems like this is not a tool that is easily to generalize, at least not in its current state. I have, and use IoT devices with API integrations, and of course I see value in a centralized platform for collating the data, but I am not even sure how I could do that with this platform.

I have no doubt the platform would work for the water monitoring as outlined in the paper, but it also seems like a lot of the code and API are specifically designed around this (and only this) use case. This includes, and is not limited to:

- Hardcoded references to basins (including API URLs, https://github.com/Applied-Artificial-Intelligence-Eurecat/acquaount/blob/main/src/main/server.js#L2231-L2258 )
- Still hardcoded IP addresses specific to the project throughout the project ( https://github.com/Applied-Artificial-Intelligence-Eurecat/acquaount/blob/main/src/main/server.js#L13-L17 )

In addition, there are many, spurious files included in the repository, a few of these are things like:

- https://github.com/Applied-Artificial-Intelligence-Eurecat/acquaount/blob/main/Error%20download%20data.PNG
- https://github.com/Applied-Artificial-Intelligence-Eurecat/acquaount/blob/main/logs_err.txt
- https://github.com/Applied-Artificial-Intelligence-Eurecat/acquaount/blob/main/git_log.txt
- https://github.com/Applied-Artificial-Intelligence-Eurecat/acquaount/blob/main/src/resources/dataIntegration/logs_err.txt

The paper states:
> The aim of WoT-Server is to provide a platform that enables easy management of time series data with minimal setup and configuration required, letting users focus on the tasks specific to their use case. With WoT-Server, developers creating IoT applications gain access to a platform where they can have their data storage up and running in minutes and use it to create AI models.

I have not found this to be the case, and there are many files that might or might not be used in configuration but it is not clearly documented as to what properties should change or have an effect.

@danielskatz I would like some advice as to how to proceed, I want to make it clear that I don't want to be critical for the sake of it, but I also would like to see in practice how the claims of the platform being able to be used in different environments. I may be way off base here, but I get the feeling that a lot of this code (including the docker files) were created with generative AI without a clear guiding structure, and made very specifically for the required setting. I also am not sure if @Ahanaf-Aziz managed to get anything working, and their last message seems somewhat AI generated (please correct me if I am wrong, and if you only used AI for translation, I apologize).

Comments like this in the code leave me feeling quite concerned:

```
// Special handling for September storage data only: replace min, max, mean with latest ENAS sensor value
        // Note: September might not appear in the endpoint response, but we update it if it exists
// ...
const septemberKey = `${septemberYear}-09-02T08:02:26.380Z`;
```

this includes about 100 lines of code dealing with "september".

I don't want to offend anyone but I also couldn't endorse September-specific code being applied in a available tool that can be deployed publicly on the internet.
