### 2026-02-06 #9470 @Ahanaf-Aziz
Hello @danielskatz  👋🏽

Thanks @zeyus , for the follow up.
After reviewing your findings , I do agree with your assessment.

I want to clarify the difference in our workflows and which eventually explains why I and even @shahchiragh initially found the software functional while you found critical errors :

### 1. Deployment vs Source Code :
I focused heavily on the user experience of getting the server running . Basically, I ran the server, requested the list of "Things" (a `GET` request), and the server successfully returned the list. This proved the server can work under perfect conditions. The Docker containers launch successfully, the server binds to `Port 2080`, and the API reads data correctly.

In contrast, @zeyus  dived deeper into the source code and edge cases. Your investigation correctly identified the blockers that the deployment tests missed . Specifically the project specific "September" logic in `server.js`, the hardcoded IPs, and the `POST/PATCH `failures.

### 2. Technical Validation :
Your Postman screenshots clarify exactly why the tests are failing :

Postman Failures : The` TypeError `in your first screenshot suggests a chain reaction failure. If an initial "Create" request fails, the script captures no ID, causing subsequent requests to crash when trying to read properties of undefined.

405 Method Not Allowed : The failures on `POST` requests imply the server is not configured to accept new data, which aligns with the hardcoding issues you found.

I completely agree with your findings on the hardcoded IP addresses, the specific "September" logic, and the junk files (`.PNG`, `logs_err.txt`). I missed these by focusing too strictly on the "Does it run?" aspect of the checklist. These definitely hinder generalizability.

### 3. Regarding my writing style  :
I used Generative AI tool to help  polishing my writing  tone from a reviewer perspective . I apologize if that made my writings  sounds like robotic .
However , please be assured that the Docker testing, curl verification, and documentation checks I reported since December were performed manually by me on my local machine.

### Updated Recommendation :
In light of the blocking issues raised by @zeyus specifically the hardcoding,  junk files, and lack of write-functionality , I am rescinding my "Approve" recommendation.

I recommend **Major Revisions**.

@Galahad3x  needs to address the issues mentioned by above to remove the project specific logic and ensure the API supports standard write operations before this can be published.
