### 2026-02-06 #9470 @danielskatz
👋 @zeyus - thanks very much for your detailed work and analysis.  Let's see what @Galahad3x says.  I see a number of issues you raise that they would need to deal with, and one question is if they can and choose to.

A larger issue might be in the last part of your message, which seems to imply that GenerativeAI was used in the code process in a way that has not produced code that is understandable, robust, general, and able to be contributed to by the larger community. Is this a fair assessment?

After this submission started review, JOSS changed its policies around AI usage as seen in https://joss.readthedocs.io/en/latest/submitting.html#ai-usage-policy, but it wouldn't be fair to judge this submission against that.  However, JOSS's policy on AI-enabled software at the time this was submitted and the review started was https://joss.readthedocs.io/en/pre-2026-submission/submitting.html#interim-position-on-ai-enabled-software, and if you (and other reviewers) feel that this submission is out-of-scope based on that policy, we should reject it as being out-of-scope, even if we didn't discover this until now.
