### 2026-02-06 #9470 @danielskatz
👋 @Ahanaf-Aziz - thanks very much for your summary and reinforcement of the issues mentioned by @zeyus. I particularly appreciate the specificity of your points.

However, just to be clear, the goal of JOSS reviews is to help the author improve the submission to where it can be accepted, if possible. Reviewers don't provide overall recommendations (such as major revisions), just judgement on which criteria are and are not met, and ideally, suggestions on how they could be met. If all the criteria are met, the review is complete. Of course, reviewers can also make additional suggestions to further improve the submission.
