### 2026-02-09 #9470 @zeyus
> A larger issue might be in the last part of your message, which seems to imply that GenerativeAI was used in the code process in a way that has not produced code that is understandable, robust, general, and able to be contributed to by the larger community. Is this a fair assessment?

That is a fair assessment, although for clarity, I think the specific issues I mentioned indicate that the product does not align with the paper and the claims made in it, regardless of whether Generative AI was used to create it or not.
