### 2026-02-09 #9470 @Galahad3x
Hello all,
I've just made some modifications, although I have not yet completed all mentioned items yet. Most of the changes in the technical validation section I still need to thoroughly test.
I have updated the README file so it is more clear in the section regarding running the init_database.py script, including the Python version, the requirements.txt file, and the script parameters. I have yet to test if this works 100%, I have some issues with Docker at the moment.
In regards to the concerns about the tool being hard to generalize, I've cleaned up most of the code specific to the use case, leaving just enough of it to be used as an example for custom development for each developer's use case. This includes hard coded references to URLs, water management/basin specific code, references to the project's pilot locations, as well as code and files related to the specific use case, mostly thing descriptions and data integration functions.
On the topic of the data integration service, I have cleaned it up and written its own README for it, so that developers can configure it with their own use case.
Related with the generalisation of the platform, I have checked to make sure there is no more "September" specific code, it was deleted along with other basin/water specific code.
Finally, I want to comment on the GenAI situation, to clarify the use of it in the project. In this project, we have used ChatGPT for small sections of code, mostly on Docker related files, but we have not used any kind of Generative AI tool for documentation (README or the paper itself). I have added an AI Usage Disclosure section at the end of the README with this same information.

In conclusion, the project is now more generalised, but it is not yet fully tested. Once it is fully tested, I will notify so the validation can be retried.
