### 2026-02-09 #9470 @Galahad3x
Hello all, I've retested the setup and deployment following the process in the README and I can assure I have been successful and the WoT-Server works as expected.
In regard to the issues mentioned by @zeyus in this comment (https://github.com/openjournals/joss-reviews/issues/9470#issuecomment-3854818564), I believe I found the cause for the errors when testing the Postman Collection. There are two main causes, the first one, as @Ahanaf-Aziz suggested, is possibly a chain failure caused by the error when running the init_database.py script. This script has been fixed to work correctly and not fail when creating the required instances in the database, even with custom URLs.
Another cause for the errors is that, as it can be seen in the screenshots, there are two "/" in most URLs where there should only be one, which causes the 405 error in the POST requests. For GET requests, it may return a 200 OK code but it's not actually returning the correct value, as it can be seen, some of the scripts in the Postman requests still fail because of it. This is probably because when configuring the environment variables, there is a trailing "/" in the URL variable which combined with the "/" in the request causes this. To fix this, make sure there is no trailing "/" in the url environment variable. I have also adjusted the documentation inside the Postman collection to notify this to future users.
