### 2026-02-16 #9470 @danielskatz
👋 @shahchiragh - I note you have a couple of unchecked items in your review. You mentioned this blocking issues in https://github.com/openjournals/joss-reviews/issues/9470#issuecomment-3764123805  And @Galahad3x tried to address this, as mentioned in https://github.com/openjournals/joss-reviews/issues/9470#issuecomment-3810692822  Can you see if this lets you complete your review? And if not, please let us know what else you think is needed.
