### 2026-02-17 #9470 @shahchiragh
Hi @danielskatz sorry for the delay. I took another look at the updated paper.md and was able to regenerate the PDF successfully. However, I do not see substantive improvements addressing my comments in https://github.com/openjournals/joss-reviews/issues/9470#issuecomment-3810692822

@Galahad3x please see my comments below :

- Statement of Need is still too weak (this needs strengthening): the current statement primarily describes the ACQUAOUNT project and what WoT-Server does. It still does not clearly articulate:
  - Who beyond ACQUAOUNT would benefit from this software?
  - Why existing solutions are insufficient?
  - What gap in the research software ecosystem this fills

While node-wot and Mozilla WebThings are mentioned there is no explicit comparison to alternative solutions such as:
   - FROST alone
   - ThingsBoard
   - FIWARE
   - Existing SensorThings-compliant servers

What is the unique research software contribution here?

- Research noveltly: Its still unclear whether WoT-server represents distinct research contribution vs. an engineered integration of existing standards? It should be position the work within broader IoT & research software landscape

- Impact: currently the article is very centered on ACQUAOUNT. Can you add some evidences on broader applicability? Or define intended audience? for instance,

> WoT-Server targets researchers and practitioners building interoperable IoT data infrastructures compliant with W3C and OGC standards.......

At this stage, I would still question following 3 categories, to finish up the review from my end:
- Research positioning (needs improvement)
- Novelty justification (weak)
- Broader impact (needs clarity)
