### 2026-02-23 #9470 @Galahad3x
@shahchiragh I have rewritten the Statement of Need and State Of The Art sections to address the questions you present in the latest comment. Let me know what you think!
