### 2026-02-23 #9470 @shahchiragh
Hi @Galahad3x sure, let me regenerate the pdfs and get back to you shortly! Thank you for addressing them promptly.
