### 2026-02-25 #9470 @zeyus
OK I have managed to run some tests again locally. It still doesn't pass.

```bash
docker-compose up -d --build
python3 src/resources/databaseInitialisation/init_database.py  src/resources/databaseInitialisation/entiredemo.csv http://localhost:8008
newman run WoT-Server\ Examples.postman_collection.json -e WOT-SERVER.postman_environment.json
```

<summary>JSON files</summary>

WoT-Server Examples.postman_collection.json:
```JSON
{
  "info": {
    "_postman_id": "1fb648f3-9815-474e-9fbf-570d0d4d4bb5",
    "name": "WoT-Server Examples",
    "description": "This folder contains example HTTP requests and examples of successful HTTP responses for the WoT-Server.\n\nThe title of the thing object (set in the thing description) is the one later used as the thing URL. Make sure this title is not prefixed by a \"/\", since this would cause errors and unexpected behaviour. To make sure these errors are not happening, make sure that all the URLs tested do not have a double \"/\" between the Thing URL and the API URL.",
    "schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
    "_exporter_id": "52142402",
    "_collection_link": "https://go.postman.co/collection/52142402-1fb648f3-9815-474e-9fbf-570d0d4d4bb5?source=collection_link"
  },
  "item": [
    {
      "name": "WoT-Server",
      "item": [
        {
          "name": "Read Field Properties",
          "request": {
            "method": "GET",
            "header": [],
            "url": {
              "raw": "{{apiURL}}/{{thing}}/properties/fieldInformation",
              "host": [
                "{{apiURL}}"
              ],
              "path": [
                "{{thing}}",
                "properties",
                "fieldInformation"
              ]
            },
            "description": "This request is used to read the Properties of a field in the WoT-Server, which contain the name, description, associated Pilot location, and coordinates."
          },
          "response": [
            {
              "name": "Read Field Properties Successful",
              "originalRequest": {
                "method": "GET",
                "header": [],
                "url": {
                  "raw": "{{apiURL}}/{{thing}}/properties/fieldInformation",
                  "host": [
                    "{{apiURL}}"
                  ],
                  "path": [
                    "{{thing}}",
                    "properties",
                    "fieldInformation"
                  ]
                }
              },
              "status": "OK",
              "code": 200,
              "_postman_previewlanguage": null,
              "header": [
                {
                  "key": "Server",
                  "value": "nginx/1.19.8"
                },
                {
                  "key": "Date",
                  "value": "Fri, 09 Jan 2026 10:29:38 GMT"
                },
                {
                  "key": "Content-Type",
                  "value": "application/json"
                },
                {
                  "key": "Transfer-Encoding",
                  "value": "chunked"
                },
                {
                  "key": "Connection",
                  "value": "keep-alive"
                },
                {
                  "key": "Access-Control-Allow-Origin",
                  "value": "*"
                }
              ],
              "cookie": [],
              "body": "{\n    \"name\": \"Demo Field\",\n    \"description\": \"Demo sensors in ABI\",\n    \"pilot\": \"Demo\",\n    \"location\": {\n        \"type\": \"Point\",\n        \"coordinates\": [\n            8.56432347,\n            40.72700569\n        ]\n    }\n}"
            }
          ]
        },
        {
          "name": "List Of Things",
          "request": {
            "method": "GET",
            "header": [],
            "url": {
              "raw": "{{apiURL}}",
              "host": [
                "{{apiURL}}"
              ]
            },
            "description": "This request returns, in a list, all the active things in the WoT-Server. The URLs returned however will not be functional, since they are prefixed with the internal IP and not the external, accessible one."
          },
          "response": [
            {
              "name": "List Of Things Successful",
              "originalRequest": {
                "method": "GET",
                "header": [],
                "url": {
                  "raw": "{{apiURL}}",
                  "host": [
                    "{{apiURL}}"
                  ]
                }
              },
              "status": "OK",
              "code": 200,
              "_postman_previewlanguage": null,
              "header": [
                {
                  "key": "Server",
                  "value": "nginx/1.19.8"
                },
                {
                  "key": "Date",
                  "value": "Fri, 09 Jan 2026 10:33:20 GMT"
                },
                {
                  "key": "Content-Type",
                  "value": "application/json"
                },
                {
                  "key": "Transfer-Encoding",
                  "value": "chunked"
                },
                {
                  "key": "Connection",
                  "value": "keep-alive"
                },
                {
                  "key": "Access-Control-Allow-Origin",
                  "value": "*"
                }
              ],
              "cookie": [],
              "body": "[\n    \"http://172.18.0.4:80/main\",\n    \"http://172.18.0.4:80/xavierc-xavitest241125-test1\",\n    \"http://172.18.0.4:80/xavierc-xavitest-v2241125-test-v2-1\",\n    \"http://172.18.0.4:80/xavierc-xavitest-guccigangguccig-gang-test-v2-guccigang\",\n    \"http://172.18.0.4:80/xavier-xavitest2-test\",\n    \"http://172.18.0.4:80/xavier-test-field-xavi-test\",\n    \"http://172.18.0.4:80/sestu-camposestutest-test\",\n    \"http://172.18.0.4:80/rob-test6-test\",\n    \"http://172.18.0.4:80/rob-test5-test\",\n    \"http://172.18.0.4:80/rob-test4-test\",\n    \"http://172.18.0.4:80/robsanpra-supercomplicatedfieldnamewowoo2-test\",\n    \"http://172.18.0.4:80/robsanpra-supercomplicatedfieldnamewowoo-test\",\n    \"http://172.18.0.4:80/robsanpra-supercomp1-test\",\n    \"http://172.18.0.4:80/robsanpra-supercomp-test\",\n    \"http://172.18.0.4:80/robsanpra-su-test\",\n    \"http://172.18.0.4:80/rob_san_prasupercomplicated_field_nametest\",\n    \"http://172.18.0.4:80/marco-campo-marco-test-test\",\n    \"http://172.18.0.4:80/marcoc-campovillasimiustest-test\",\n    \"http://172.18.0.4:80/marcoc-campomarcotest-test\",\n    \"http://172.18.0.4:80/edc68a17-campoarborea-test\",\n    \"http://172.18.0.4:80/ec74d4a1-arborea-test\",\n    \"http://172.18.0.4:80/e9b4b55b-champ1-test\",\n    \"http://172.18.0.4:80/e564de0f-namefield-test\",\n    \"http://172.18.0.4:80/dfd150f0-campotest-test\",\n    \"http://172.18.0.4:80/da91812a-lebanonfieldtest-test\",\n    \"http://172.18.0.4:80/d69a1268-campoussana-test\",\n    \"http://172.18.0.4:80/cbd2f135-tunisianworkshop-test\",\n    \"http://172.18.0.4:80/cb3bc6fb-potato-test\",\n    \"http://172.18.0.4:80/c721853f-campovillasimius2-test\",\n    \"http://172.18.0.4:80/bc14b19e-newfieldpisa-test\",\n    \"http://172.18.0.4:80/bacf58f1-camposestu-test\",\n    \"http://172.18.0.4:80/andrecampoandreatesttest\",\n    \"http://172.18.0.4:80/ad1f3f79-campocastiadas-test\",\n    \"http://172.18.0.4:80/acquaountjohndoemynewfieldtesting\",\n    \"http://172.18.0.4:80/acquaount-pivettabasso-prod\",\n    \"http://172.18.0.4:80/ab8a4eaf-champ2-test\",\n    \"http://172.18.0.4:80/ab31nx81-campovillasimiustest-test\",\n    \"http://172.18.0.4:80/a947d949-campovilla2-test\",\n    \"http://172.18.0.4:80/a155d1e7-olivetooristano-test\",\n    \"http://172.18.0.4:80/9eae899c-bednayel-test\",\n    \"http://172.18.0.4:80/98dd1088-centroaziendalenosensori-test\",\n    \"http://172.18.0.4:80/9629d0eb-campovillasimius-test\",\n    \"http://172.18.0.4:80/9219397a-campovilla1-test\",\n    \"http://172.18.0.4:80/5b899e48-camposestu-test\",\n    \"http://172.18.0.4:80/5790fa4a-faisal-test\",\n    \"http://172.18.0.4:80/444aa653-uliveto-test\",\n    \"http://172.18.0.4:80/3c2ed71e-fieldbarcelona-test\",\n    \"http://172.18.0.4:80/2a5334f6-campovillasimius-test\",\n    \"http://172.18.0.4:80/1d0f1363-karamah-test\",\n    \"http://172.18.0.4:80/139086b9-aloui-test\",\n    \"http://172.18.0.4:80/131-campovillasimiustest-test\",\n    \"http://172.18.0.4:80/131-campovillasimius-test\",\n    \"http://172.18.0.4:80/131-campodolianova-test\",\n    \"http://172.18.0.4:80/0b5e3f64-casa-test\",\n    \"http://172.18.0.4:80/090f31d8-campocastiadas-test\",\n    \"http://172.18.0.4:80/053695f2-shhab-test\",\n    \"http://172.18.0.4:80/acquaounttorresan\",\n    \"http://172.18.0.4:80/acquaountpivetta\",\n    \"http://172.18.0.4:80/acquaountpinos\",\n    \"http://172.18.0.4:80/acquaountpinnacentroaziendale\",\n    \"http://172.18.0.4:80/acquaountpinnacamedda\",\n    \"http://172.18.0.4:80/acquaountdesogusperra2\",\n    \"http://172.18.0.4:80/acquaountdesogusmazzacaddeo\",\n    \"http://172.18.0.4:80/acquaountyahyaabdulrahmandoganabuobaideh\",\n    \"http://172.18.0.4:80/acquaounttelalalthahab\",\n    \"http://172.18.0.4:80/acquaountnarc\",\n    \"http://172.18.0.4:80/acquaountmuflehal-sawalha\",\n    \"http://172.18.0.4:80/acquaountmahmoudal-rajbi\",\n    \"http://172.18.0.4:80/acquaountkhalilal-edwan\",\n    \"http://172.18.0.4:80/acquaountibrahimal-shareef\",\n    \"http://172.18.0.4:80/acquaountfaisalneimat\",\n    \"http://172.18.0.4:80/acquaountes-haqsawalha\",\n    \"http://172.18.0.4:80/acquaountatifabdelrahim\",\n    \"http://172.18.0.4:80/acquaountatefal-adi\",\n    \"http://172.18.0.4:80/acquaountsaraainbaalbeck\",\n    \"http://172.18.0.4:80/acquaountsaaidehbaalbeck\",\n    \"http://172.18.0.4:80/acquaountjaoudatelhajyoussef\",\n    \"http://172.18.0.4:80/acquaounthusseinfayad\",\n    \"http://172.18.0.4:80/acquaounthaouchelsaaloukwestbekaa\",\n    \"http://172.18.0.4:80/acquaountghabibshara\",\n    \"http://172.18.0.4:80/acquaountfadikhoury\",\n    \"http://172.18.0.4:80/acquaountabdallahhanna\",\n    \"http://172.18.0.4:80/acquaountabbassabdallah\",\n    \"http://172.18.0.4:80/acquaountsalemaloui\",\n    \"http://172.18.0.4:80/acquaountnoureddinerojbanifarm\",\n    \"http://172.18.0.4:80/acquaountmohamedbouzidi\",\n    \"http://172.18.0.4:80/acquaountmabroukacompany\",\n    \"http://172.18.0.4:80/acquaountlaminemoussa\",\n    \"http://172.18.0.4:80/acquaountjomaadhibi\",\n    \"http://172.18.0.4:80/acquaountthingdemosite\",\n    \"http://172.18.0.4:80/acquaounttirsoaquifer\",\n    \"http://172.18.0.4:80/acquaountcantonierareservoir\",\n    \"http://172.18.0.4:80/acquaounttriassicacquifer\",\n    \"http://172.18.0.4:80/acquaountyammounehlake\",\n    \"http://172.18.0.4:80/acquaountoldmatmata\",\n    \"http://172.18.0.4:80/acquaountnoureddinerojbani\",\n    \"http://172.18.0.4:80/acquaountmatmatanouvelle\",\n    \"http://172.18.0.4:80/acquaountjordanks\",\n    \"http://172.18.0.4:80/acquaountjordanda\",\n    \"http://172.18.0.4:80/acquaountlebanon00204eb1\",\n    \"http://172.18.0.4:80/acquaountlebanon00204ea1\",\n    \"http://172.18.0.4:80/acquaountlebanon00204b21\",\n    \"http://172.18.0.4:80/acquaountlebanon000025bd\",\n    \"http://172.18.0.4:80/acquaountlebanon00001896\",\n    \"http://172.18.0.4:80/acquaountlebanon000010cc\",\n    \"http://172.18.0.4:80/acquaountlebanon000008f7\",\n    \"http://172.18.0.4:80/acquaountlebanon000008e5\",\n    \"http://172.18.0.4:80/acquaountlebanon000008e4\",\n    \"http://172.18.0.4:80/acquaountlebanon000008e0\",\n    \"http://172.18.0.4:80/acquaountlebanon000008dd\",\n    \"http://172.18.0.4:80/acquaountlebanon000008d6\",\n    \"http://172.18.0.4:80/acquaountdesogus\"\n]"
            }
          ]
        },
        {
          "name": "Read Sensor List",
          "request": {
            "method": "GET",
            "header": [],
            "url": {
              "raw": "{{apiURL}}/{{thing}}/properties/sensorsList",
              "host": [
                "{{apiURL}}"
              ],
              "path": [
                "{{thing}}",
                "properties",
                "sensorsList"
              ]
            },
            "description": "This request will return a list of all the sensors associated with the specified thing. It will also return internal links to related datastreams in the SensorThings server."
          },
          "response": [
            {
              "name": "Read Sensor List Succesful",
              "originalRequest": {
                "method": "GET",
                "header": [],
                "url": {
                  "raw": "{{apiURL}}/{{thing}}/properties/sensorsList",
                  "host": [
                    "{{apiURL}}"
                  ],
                  "path": [
                    "{{thing}}",
                    "properties",
                    "sensorsList"
                  ]
                }
              },
              "status": "OK",
              "code": 200,
              "_postman_previewlanguage": null,
              "header": [
                {
                  "key": "Server",
                  "value": "nginx/1.19.8"
                },
                {
                  "key": "Date",
                  "value": "Fri, 09 Jan 2026 10:36:21 GMT"
                },
                {
                  "key": "Content-Type",
                  "value": "application/json"
                },
                {
                  "key": "Transfer-Encoding",
                  "value": "chunked"
                },
                {
                  "key": "Connection",
                  "value": "keep-alive"
                },
                {
                  "key": "Access-Control-Allow-Origin",
                  "value": "*"
                }
              ],
              "cookie": [],
              "body": "[\n    {\n        \"@iot.selfLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(1)\",\n        \"@iot.id\": 1,\n        \"name\": \"D001\",\n        \"description\": \"Compact and versatile environmental sensor module designed for monitoring temperature and humidity in various applications.\",\n        \"encodingType\": \"application/pdf\",\n        \"metadata\": \"Compact and versatile environmental sensor module designed for monitoring temperature and humidity in various applications.\",\n        \"properties\": {\n            \"deviceEUI\": \"2cf7f1c043500183\",\n            \"deviceType\": \"ST01\"\n        },\n        \"Datastreams@iot.navigationLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(1)/Datastreams\",\n        \"MultiDatastreams@iot.navigationLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(1)/MultiDatastreams\"\n    },\n    {\n        \"@iot.selfLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(2)\",\n        \"@iot.id\": 2,\n        \"name\": \"D002\",\n        \"description\": \"A sensor module for measuring soil moisture and temperature, ideal for agriculture and environmental monitoring.\",\n        \"encodingType\": \"application/pdf\",\n        \"metadata\": \"A sensor module for measuring soil moisture and temperature, ideal for agriculture and environmental monitoring.\",\n        \"properties\": {\n            \"deviceEUI\": \"2CF7F12031500017\",\n            \"deviceType\": \"ST02\"\n        },\n        \"Datastreams@iot.navigationLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(2)/Datastreams\",\n        \"MultiDatastreams@iot.navigationLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(2)/MultiDatastreams\"\n    },\n    {\n        \"@iot.selfLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(3)\",\n        \"@iot.id\": 3,\n        \"name\": \"D003\",\n        \"description\": \"A LoRa-enabled controller for managing and controlling IoT devices, facilitating long-range wireless communication.\",\n        \"encodingType\": \"application/pdf\",\n        \"metadata\": \"A LoRa-enabled controller for managing and controlling IoT devices, facilitating long-range wireless communication.\",\n        \"properties\": {\n            \"deviceEUI\": \"24e124414b106165\",\n            \"deviceType\": \"ST03\"\n        },\n        \"Datastreams@iot.navigationLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(3)/Datastreams\",\n        \"MultiDatastreams@iot.navigationLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(3)/MultiDatastreams\"\n    },\n    {\n        \"@iot.selfLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(4)\",\n        \"@iot.id\": 4,\n        \"name\": \"D004\",\n        \"description\": \"An advanced sensor capable of monitoring temperature and moisture at different depths. Useful for agronomic research activities.\",\n        \"encodingType\": \"application/pdf\",\n        \"metadata\": \"An advanced sensor capable of monitoring temperature and moisture at different depths. Useful for agronomic research activities.\",\n        \"properties\": {\n            \"deviceEUI\": \"3,73838E+15\",\n            \"deviceType\": \"ST04\"\n        },\n        \"Datastreams@iot.navigationLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(4)/Datastreams\",\n        \"MultiDatastreams@iot.navigationLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(4)/MultiDatastreams\"\n    },\n    {\n        \"@iot.selfLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(5)\",\n        \"@iot.id\": 5,\n        \"name\": \"D005\",\n        \"description\": \"A sensor for counting and transmitting the pulse produced by a flowmeter.\",\n        \"encodingType\": \"application/pdf\",\n        \"metadata\": \"A sensor for counting and transmitting the pulse produced by a flowmeter.\",\n        \"properties\": {\n            \"deviceEUI\": \"AC1F09FFFE04E43F \",\n            \"deviceType\": \"ST05\"\n        },\n        \"Datastreams@iot.navigationLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(5)/Datastreams\",\n        \"MultiDatastreams@iot.navigationLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(5)/MultiDatastreams\"\n    },\n    {\n        \"@iot.selfLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(186)\",\n        \"@iot.id\": 186,\n        \"name\": \"demo field\",\n        \"description\": \"Model outputs for a field\",\n        \"encodingType\": \"application/pdf\",\n        \"metadata\": \"Sensor\",\n        \"properties\": {},\n        \"Datastreams@iot.navigationLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(186)/Datastreams\",\n        \"MultiDatastreams@iot.navigationLink\": \"http://localhost:8008/FROST-Server/v1.1/Sensors(186)/MultiDatastreams\"\n    }\n]"
            }
          ]
        },
        {
          "name": "Read Sensor Information",
          "request": {
            "method": "GET",
            "header": [],
            "url": {
              "raw": "{{apiURL}}/{{thing}}/properties/sensorInformation?deviceID=D001",
              "host": [
                "{{apiURL}}"
              ],
              "path": [
                "{{thing}}",
                "properties",
                "sensorInformation"
              ],
              "query": [
                {
                  "key": "deviceID",
                  "value": "D001"
                }
              ]
            },
            "description": "This request will return the information of one specific sensor, defined in the query parameters. It will only be successful if the sensor both exists in the database and is associated with the specified thing."
          },
          "response": [
            {
              "name": "Read Sensor Information Successful",
              "originalRequest": {
                "method": "GET",
                "header": [],
                "url": {
[…]
