### 2026-02-26 #9470 @Galahad3x
Hello @zeyus, I see some variables in the Postman collection from the specific ACQUAOUNT use case such as those you mention slipped past me when I was making changes to the repository to generalise the project. I have changed them now to use the examples provided in the demo files used with the init_database.py file. While the string variables have been fixed, it is possible that some requests still show unexpected behaviour because of the use of numerical IDs which can change depending on the order of the requests or the example file used to test. Keep in mind that the purpose of the Postman collection is to show users and developers examples of how to perform requests to the WoT-Server and not being a set of automated tests for the service, so they might fail without human intervention to assure the correct IDs or variables are being used.
Regarding the API documentation, the Postman Collection serves as the documentation together with the list of interaction affordances in the README and the base Thing Description itself, I have updated the README to link to the Thing Description.
