### 2026-03-03 #9470 @zeyus
hi @danielskatz yes, I have marked what I can on the checklist.

I think though the remaining points:

- Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I don't think that a list of endpoints is API or functional documentation, but if this is acceptable, I will check the box

- Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Requiring a postman account and running requests against and endpoint is not automated testing

Apart from that, the others are satisfactory.
