### 2026-03-03 #9470 @danielskatz
👋 @shahchiragh - Have @Galahad3x's [recent changes](https://github.com/openjournals/joss-reviews/issues/9470#issuecomment-3943715740) moved the submission to the point where you can check off your final items?
