### 2026-03-03 #9470 @danielskatz
👋 @Galahad3x - please let us know how you think we can resolve [the issues](https://github.com/openjournals/joss-reviews/issues/9470#issuecomment-3991019162) that @zeyus still has?
