### 2026-03-09 #9470 @Galahad3x
Hello @danielskatz, I have added API documentation as a Markdown file in the root directory named api.md and implemented automatic testing using pytest to check the last issues!
