### 2026-03-09 #9470 @zeyus
> Hello [@danielskatz](https://github.com/danielskatz), I have added API documentation as a Markdown file in the root directory named api.md and implemented automatic testing using pytest to check the last issues!

Hi @Galahad3x thanks for that I will take a look!
