### 2026-03-10 #9470 @shahchiragh
Hi @Galahad3x, thanks for considering the suggestions. Upon further reading, the paper partially addresses the concerns. The State of the Art section provides some comparisons to alternative solutions such as FROST-Server, ThingsBoard, and FIWARE, stating that WoT-Server is among the first open-source platforms to combine the Web of Things interaction model with a SensorThings-compliant data model and persistent storage .

However, the Statement of Need still primarily focuses on the ACQUAOUNT project. It does mention that the challenges addressed by WoT-Server are common across many research domains, such as agriculture, environmental monitoring, hydrology, and smart cities, indicating a broader applicability.

Regarding research novelty, the document highlights that WoT-Server combines the Web of Things and SensorThings standards, providing a unique contribution. The aim is to provide a platform with minimal setup, aligned to standards, and with a centralized API that enables easy management of time series data for research-grade IoT data management.

To address the remaining concerns, the document could still further clarify the following:
- Who beyond ACQUAOUNT would benefit from WoT-Server
- Why existing solutions like FROST alone, ThingsBoard, and FIWARE are insufficient
- The unique research software contribution of WoT-Server
- The broader impact and intended audience of WoT-Server

A potential statement to clarify the intended audience could be: "WoT-Server targets researchers and practitioners building interoperable IoT data infrastructures compliant with W3C and OGC standards."

While the document provides some insights into the broader applicability and research positioning of WoT-Server, further clarification is still needed to fully address the concerns.
