### 2026-03-10 #9470 @shahchiragh
Hi @danielskatz; I've outlined my concerns above and would like to seek further clarification on the "Statement of Need" and "State of the field" sections. While the references appear to be in order, I'm still uncertain about the weightage assigned to these two sections.

To move forward, I would appreciate it if someone could review the comparisons made in the paper to provide a final confirmation before giving the green light.

Thanks,
Chirag
