### 2026-03-10 #9470 @danielskatz
👋 @shahchiragh - I think that your clarification requests are mostly reasonable. Since JOSS doesn't require that software be unique, I would drop the third one, but I think that if @Galahad3x were to address

* Who beyond ACQUAOUNT would benefit from WoT-Server
* Why existing solutions like FROST alone, ThingsBoard, and FIWARE are insufficient
* The broader impact and intended audience of WoT-Server (where you have already suggested a sentence above)

by adding some text to the Statement of Need section, I personally think this would be sufficient. Would this work for you?
