### 2026-03-11 #9470 @Galahad3x
Hello @danielskatz and @shahchiragh, I have added the statement you proposed to the Statement Of Need section, but I wanted to ask a few questions about the remaining concerns.
I believe the first one and the third one to be solved also with the same statement, describing that researchers for projects of many different domains would benefit from WoT-Server, not just the ACQUAOUNT Project.
As to why existing solutions like FROST, ThingsBoard, FIWARE, and others are insufficient for these use cases, this is described in the State Of The Art section. Should it stay there or should this be merged with the Statement Of Need section? I don't see the value in having the duplicated information in both sections.
