### 2026-03-11 #9470 @danielskatz
@Galahad3x - under the JOSS guidelines in place when this submission started, there is no requirement for a State of the Art section, so I think it would be fine for these two to be combined as Statement of Need. I hope this is ok with @shahchiragh as well.
