### 2026-03-11 #9470 @Ahanaf-Aziz
Hi @danielskatz  and @Galahad3x ,
I'm happy to confirm that I have completed my review of this submission. It has been a genuinely enjoyable and rewarding experience to be part of this entire review process.

When I first started the review,  there were several blocking issues like  from installation documentation, hardcoded URLs, missing tests, and incomplete references. I'm pleased to see how @Galahad3x responded to every concern promptly and  making meaningful improvements at each stage.

The final version of the repository and paper is significantly better than when we started and now we observe a cleaner codebase , proper documentation, automated tests, community guidelines, and a well positioned State of the Art section.

I'm satisfied with the current state of the submission and happy to recommend it for publication.

Thank you to @danielskatz  for keeping the process on track and to @Galahad3x  for their patience and dedication throughout.

I would be happy to collaborate again in the future !

Best regards

Ahanaf
