### 2026-03-23 #9470 @danielskatz
👋 @shahchiragh - you still have one unchecked item in [your checklist](https://github.com/openjournals/joss-reviews/issues/9470#issuecomment-3572394157). Does the discussion and changes in the 3 or 4 issues above allow you to check this item now?
