### 2026-03-23 #9470 @shahchiragh
@danielskatz Apologies again for the delay. I've completed my review of the revised document and it looks good overall. I'd like to acknowledge @Galahad3x for the timely and thorough updates; the submission now addresses all items on the checklist. It's been a great experience so far. Let me know if there's any pending items from my end.

Thanks,
Chirag
