### 2026-03-24 #9470 @Galahad3x
That is great news! I am happy that the current version of the software meets certain standards of quality.

@danielskatz What are the next steps?
