### 2026-03-24 #9470 @Galahad3x
Hello @danielskatz, yes, it looks like the DOI is correct
