### 2026-03-24 #9470 @danielskatz
👋 @Galahad3x - I've suggested a few minor changes in https://github.com/Applied-Artificial-Intelligence-Eurecat/acquaount/pull/2 including adding the missing DOI.  Please merge this, or let me know what you disagree with, then we can continue.
