### 2026-03-26 #9470 @danielskatz
👋 @Galahad3x - the problem might be that dois are allowed in @online entries, so I'm trying to change it to @misc.  Can you see if merging https://github.com/Applied-Artificial-Intelligence-Eurecat/acquaount/pull/3 lets this move forward?
