### 2026-03-30 #9470 @danielskatz
@Galahad3x - ok, now I see that I introduced mismatched brackets - please merge https://github.com/Applied-Artificial-Intelligence-Eurecat/acquaount/pull/4 to fix this...
