### 2026-04-07 #9470 @Galahad3x
Hello @danielskatz, I have completed the tasks in the list:

- A tagged version of the software is now created, v2.0
- This version has been uploaded to Zenodo, with DOI https://doi.org/10.5281/zenodo.19450125
