### 2026-04-07 #9470 @danielskatz
Congratulations to @Galahad3x (Joel Aumedes Serrano) and co-authors on your JOSS publication!!

And thanks to @zeyus, @Ahanaf-Aziz, and @shahchiragh for reviewing!

JOSS completely depends on volunteers - we couldn't do this without you
